Make INotify be more @async friendly
Calling into a C function that blocks, such as ccalling the system
`read()` function, is bad for Julia because that thread blocks.  This
effectively blocks all async Tasks that re running on the same thread.
To be more friendly to such Tasks, we now use inotify with `O_NONBLOCK`
so that reads will not block.  To emulate blocking reads, we now convert
the file descriptor returned by Julia into a `RawFD`.  `RawFD` values
can be passed to C functions expecting an integer file descriptor.  They
can also be passed to Julia's `wait()` function to block in a
Julia-aware and asynchronous-friendly way.