Give each inotify instance a sentinal file
By giving each inotify instance a known sentinal file and setting up an
watch on the inotify instance's sentinal file, it is possible to
generate events on any (or all) inotify instances.  This can be useful
for unblocking Tasks that may be blocked in a system `read()` call on
the inotify file descriptor.  A new `inotify_notify` function does just
that.

The looping in the `inotify_read_events` methods has also been
eliminated.  Of course, users can write their own loops if desired.