Make DirWatcher robust against race condition
When creating multiple levels of subdirectories in a watched directory,
it is possible for the second level subdirectory to be created before
DirWatcher has added a watch to the first level subdirectory.  Prior to
this commit, DirWatcher would be oblivious to the second level
subdirectory so any/all activity there would be missed.  To avoid this
predicament, whenever DirWatcher adds a watch to a newly created
directory it walks the contents of that newly created directory to
ensure that it will detect any subdirectories or files that were created
before the watch was added.  It is possible that this will detetct files
created immediately after the watch was created which will also be
detected by the watch, so clients should be prepared to handle duplicate
creation events for a given entity.